pyproject.toml
[project]
name = "textbookreview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.2.14",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
